LayoutTests/webarchive/loading/test-loading-archive-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - willCloseFrame
main frame - didCommitLoadForFrame
main frame - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
resources/helloworld.webarchive - willSendRequest <NSURLRequest resources/helloworld.webarchive> redirectResponse (null)
main frame - didFinishDocumentLoadForFrame
<unknown> - didFinishLoading
resources/helloworld.webarchive - didReceiveResponse <NSURLResponse resources/helloworld.webarchive>
frame "<!--framePath //<!--frame0-->-->" - willCloseFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didClearWindowObjectForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFirstLayoutInFrame
main frame - didFirstLayoutInFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
 This tests that doing a "normal load" of a webarchive (not using loadArchive) does not cause a cancelled error to be called.
